+++ pkg/_fe_analyzer_shared/lib/src/messages/codes_generated.dart
@@ -7381,6 +7381,38 @@ const MessageCode messageJsInteropNonExternalMember = const MessageCode(
     correctionMessage: r"""Try annotating the member with `external`.""");
 
 // DO NOT EDIT. THIS FILE IS GENERATED. SEE TOP OF FILE.
+const Template<Message Function(String name, String name2)>
+    templateJsInteropNonStaticWithStaticInteropSupertype =
+    const Template<Message Function(String name, String name2)>(
+        problemMessageTemplate:
+            r"""Class '#name' does not have an `@staticInterop` annotation, but has supertype '#name2', which does.""",
+        correctionMessageTemplate:
+            r"""Try marking '#name' as a `@staticInterop` class, or don't inherit '#name2'.""",
+        withArguments:
+            _withArgumentsJsInteropNonStaticWithStaticInteropSupertype);
+
+// DO NOT EDIT. THIS FILE IS GENERATED. SEE TOP OF FILE.
+const Code<Message Function(String name, String name2)>
+    codeJsInteropNonStaticWithStaticInteropSupertype =
+    const Code<Message Function(String name, String name2)>(
+  "JsInteropNonStaticWithStaticInteropSupertype",
+);
+
+// DO NOT EDIT. THIS FILE IS GENERATED. SEE TOP OF FILE.
+Message _withArgumentsJsInteropNonStaticWithStaticInteropSupertype(
+    String name, String name2) {
+  if (name.isEmpty) throw 'No name provided';
+  name = demangleMixinApplicationName(name);
+  if (name2.isEmpty) throw 'No name provided';
+  name2 = demangleMixinApplicationName(name2);
+  return new Message(codeJsInteropNonStaticWithStaticInteropSupertype,
+      problemMessage:
+          """Class '${name}' does not have an `@staticInterop` annotation, but has supertype '${name2}', which does.""",
+      correctionMessage: """Try marking '${name}' as a `@staticInterop` class, or don't inherit '${name2}'.""",
+      arguments: {'name': name, 'name2': name2});
+}
+
+// DO NOT EDIT. THIS FILE IS GENERATED. SEE TOP OF FILE.
 const Code<Null> codeJsInteropOperatorsNotSupported =
     messageJsInteropOperatorsNotSupported;
 
@@ -7637,7 +7669,7 @@ const Template<Message Function(String name, String name2)>
     templateJsInteropStaticInteropWithNonStaticSupertype =
     const Template<Message Function(String name, String name2)>(
         problemMessageTemplate:
-            r"""JS interop class '#name' has an `@staticInterop` annotation, but has supertype '#name2', which is non-static.""",
+            r"""JS interop class '#name' has an `@staticInterop` annotation, but has supertype '#name2', which does not.""",
         correctionMessageTemplate:
             r"""Try marking the supertype as a static interop class using `@staticInterop`.""",
         withArguments:
@@ -7659,7 +7691,7 @@ Message _withArgumentsJsInteropStaticInteropWithNonStaticSupertype(
   name2 = demangleMixinApplicationName(name2);
   return new Message(codeJsInteropStaticInteropWithNonStaticSupertype,
       problemMessage:
-          """JS interop class '${name}' has an `@staticInterop` annotation, but has supertype '${name2}', which is non-static.""",
+          """JS interop class '${name}' has an `@staticInterop` annotation, but has supertype '${name2}', which does not.""",
       correctionMessage: """Try marking the supertype as a static interop class using `@staticInterop`.""",
       arguments: {'name': name, 'name2': name2});
 }